+++ core/domain/beam_job_domain_test.py
@@ -23,7 +23,7 @@ import datetime
 
 from core import utils
 from core.domain import beam_job_domain
-from core.jobs.batch_jobs import validation_jobs
+from core.jobs.batch_jobs import model_validation_jobs
 from core.platform import models
 from core.tests import test_utils
 
@@ -39,7 +39,8 @@ class BeamJobTests(test_utils.TestBase):
     NOW = datetime.datetime.utcnow()
 
     def test_usage(self) -> None:
-        job = beam_job_domain.BeamJob(validation_jobs.AuditAllStorageModelsJob)
+        job = beam_job_domain.BeamJob(
+            model_validation_jobs.AuditAllStorageModelsJob)
         self.assertEqual(job.name, 'AuditAllStorageModelsJob')
 
     def test_in_terminal_state(self) -> None:
@@ -90,7 +91,8 @@ class BeamJobTests(test_utils.TestBase):
         self.assertFalse(unknown_beam_job_run.in_terminal_state)
 
     def test_to_dict(self) -> None:
-        job = beam_job_domain.BeamJob(validation_jobs.AuditAllStorageModelsJob)
+        job = beam_job_domain.BeamJob(
+            model_validation_jobs.AuditAllStorageModelsJob)
         self.assertEqual(job.to_dict(), {'name': 'AuditAllStorageModelsJob'})
