+++ core/jobs/base_jobs.py
@@ -117,7 +117,11 @@ class JobMetaclass(type):
 
         if not name.endswith('Base'):
             if issubclass(job_cls, JobBase):
+                if not name.endswith('Job'):
+                    raise TypeError('Job name "%s" must end with "Job"' % name)
+
                 cls._JOB_REGISTRY[name] = job_cls
+
             else:
                 raise TypeError('%s must inherit from JobBase' % name)
