+++ core/jobs/batch_jobs/blog_validation_jobs.py
@@ -28,50 +28,18 @@ from core.platform import models
 import apache_beam as beam
 
 MYPY = False
-if MYPY: # pragma: no cover
+if MYPY:  # pragma: no cover
     from mypy_imports import blog_models
 
-(blog_models,) = models.Registry.import_models(
-    [models.NAMES.blog, ])
+(blog_models,) = models.Registry.import_models([models.NAMES.blog])
 
 
-class GetModelsWithDuplicatePropertyValues(beam.PTransform):
-    """Helper class to retrive models with duplicate properties."""
-
-    def __init__(self, property_name: str) -> None:
-        super(GetModelsWithDuplicatePropertyValues, self).__init__()
-        self._property_name = property_name
-
-    def expand(self, blog_model_pcoll):
-        return (
-            blog_model_pcoll
-            | 'Discard models with empty property value' >> (
-                beam.Filter(lambda model: self.get_property_value(model) != ''))
-            | 'Generate (%s, model) key value pairs' % self._property_name >> (
-                beam.WithKeys(self.get_property_value)) # pylint: disable=no-value-for-parameter
-            | 'Group pairs by their %s' % self._property_name >> (
-                beam.GroupByKey())
-            | 'Discard %s key' % self._property_name >> beam.Values() # pylint: disable=no-value-for-parameter
-            | 'Discard models with unique %s' % self._property_name >> (
-                beam.Filter(lambda models: len(models) > 1))
-        )
-
-    def get_property_value(self, model):
-        """Returns value of the given property of model
-
-        Args:
-            model: datastore_services.Model. Entity to validate.
-
-        Returns:
-            value. The value of the property of model.
-        """
-        return job_utils.get_model_property(model, self._property_name)
-
-
-class BlogPostTitleUniquenessJob(base_jobs.JobBase):
+class FindDuplicateBlogPostTitlesJob(base_jobs.JobBase):
     """Validates that all the Blog Posts have unique title."""
 
-    def run(self) -> beam.Pipeline:
+    def run(
+        self
+    ) -> beam.PCollection[blog_validation_errors.DuplicateBlogTitleError]:
         return (
             self.pipeline
             | 'Get every Blog Model' >> (
@@ -85,15 +53,16 @@ class BlogPostTitleUniquenessJob(base_jobs.JobBase):
         )
 
 
-class BlogPostUrlUniquenessJob(base_jobs.JobBase):
+class FindDuplicateBlogPostUrlsJob(base_jobs.JobBase):
     """Validates that all the Blog Posts have unique url."""
 
-    def run(self) -> beam.Pipeline:
+    def run(
+        self
+    ) -> beam.PCollection[blog_validation_errors.DuplicateBlogUrlError]:
         return (
             self.pipeline
             | 'Get every Blog Post Model' >> (
-                ndb_io.GetModels(
-                    blog_models.BlogPostModel.query()))
+                ndb_io.GetModels(blog_models.BlogPostModel.query()))
             | GetModelsWithDuplicatePropertyValues('url_fragment')
             | 'Flatten models into a list of errors' >> beam.FlatMap(
                 lambda models: [
@@ -103,15 +72,16 @@ class BlogPostUrlUniquenessJob(base_jobs.JobBase):
         )
 
 
-class BlogPostSummaryTitleUniquenessJob(base_jobs.JobBase):
+class FindDuplicateBlogPostSummaryTitlesJob(base_jobs.JobBase):
     """Validates that all the Blog Post Summary Model have unique title."""
 
-    def run(self) -> beam.Pipeline:
+    def run(
+        self
+    ) -> beam.PCollection[blog_validation_errors.DuplicateBlogTitleError]:
         return (
             self.pipeline
             | 'Get every Blog Summary Model' >> (
-                ndb_io.GetModels(
-                    blog_models.BlogPostSummaryModel.query()))
+                ndb_io.GetModels(blog_models.BlogPostSummaryModel.query()))
             | GetModelsWithDuplicatePropertyValues('title')
             | 'Flatten models into a list of errors' >> beam.FlatMap(
                 lambda models: [
@@ -121,15 +91,16 @@ class BlogPostSummaryTitleUniquenessJob(base_jobs.JobBase):
         )
 
 
-class BlogPostSummaryUrlUniquenessJob(base_jobs.JobBase):
+class FindDuplicateBlogPostSummaryUrlsJob(base_jobs.JobBase):
     """Validates that all the Blog Post Summary Model have unique url."""
 
-    def run(self) -> beam.Pipeline:
+    def run(
+        self
+    ) -> beam.PCollection[blog_validation_errors.DuplicateBlogUrlError]:
         return (
             self.pipeline
             | 'Get every Blog Post Summary Model' >> (
-                ndb_io.GetModels(
-                    blog_models.BlogPostSummaryModel.query()))
+                ndb_io.GetModels(blog_models.BlogPostSummaryModel.query()))
             | GetModelsWithDuplicatePropertyValues('url_fragment')
             | 'Flatten models into a list of errors' >> beam.FlatMap(
                 lambda models: [
@@ -137,3 +108,37 @@ class BlogPostSummaryUrlUniquenessJob(base_jobs.JobBase):
                     for model in models
                 ])
         )
+
+
+class GetModelsWithDuplicatePropertyValues(beam.PTransform):
+    """Helper class to retrive models with duplicate properties."""
+
+    def __init__(self, property_name: str) -> None:
+        super(GetModelsWithDuplicatePropertyValues, self).__init__()
+        self.property_name = property_name
+
+    def expand(self, blog_model_pcoll):
+        return (
+            blog_model_pcoll
+            | 'Discard models with empty property value' >> (
+                beam.Filter(lambda model: self.get_property_value(model) != ''))
+            | 'Generate (%s, model) key value pairs' % self.property_name >> (
+                beam.WithKeys(self.get_property_value)) # pylint: disable=no-value-for-parameter
+            | 'Group pairs by their %s' % self.property_name >> (
+                beam.GroupByKey())
+            | 'Discard %s key' % self.property_name >> (
+                beam.Values()) # pylint: disable=no-value-for-parameter
+            | 'Discard models with unique %s' % self.property_name >> (
+                beam.Filter(lambda models: len(models) > 1))
+        )
+
+    def get_property_value(self, model):
+        """Returns value of the given property of model
+
+        Args:
+            model: datastore_services.Model. Entity to validate.
+
+        Returns:
+            value. The value of the property of model.
+        """
+        return job_utils.get_model_property(model, self.property_name)
