+++ core/jobs/batch_jobs/model_validation_jobs_test.py
@@ -14,14 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unit tests for jobs.batch_jobs.validation_jobs."""
+"""Unit tests for jobs.batch_jobs.model_validation_jobs."""
 
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from core import feconf
 from core.jobs import job_test_utils
-from core.jobs.batch_jobs import validation_jobs
+from core.jobs.batch_jobs import model_validation_jobs
 from core.jobs.transforms.validation import base_validation
 from core.jobs.types import base_validation_errors
 from core.jobs.types import model_property
@@ -33,7 +33,7 @@ from core.platform import models
 
 class AuditAllStorageModelsJobTests(job_test_utils.JobTestBase):
 
-    JOB_CLASS = validation_jobs.AuditAllStorageModelsJob
+    JOB_CLASS = model_validation_jobs.AuditAllStorageModelsJob
 
     VALID_USER_ID = 'uid_%s' % ('a' * feconf.USER_ID_RANDOM_PART_LENGTH)
