+++ core/jobs/registry.py
@@ -44,10 +44,14 @@ from typing import List, Type
 # We need this to happen for every job in this registry file, because the
 # registry depends on JobMetaclass to handle the responsibility of keeping track
 # of every job.
-from core.jobs.batch_jobs import blog_validation_jobs        # pylint: disable=unused-import  # isort: skip
-from core.jobs.batch_jobs import cron_jobs                   # pylint: disable=unused-import  # isort: skip
-from core.jobs.batch_jobs import opportunity_management_jobs # pylint: disable=unused-import  # isort: skip
-from core.jobs.batch_jobs import validation_jobs             # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import blog_validation_jobs              # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import (                                 # pylint: disable=unused-import  # isort: skip
+    exp_recommendation_computation_jobs)
+from core.jobs.batch_jobs import exp_search_indexing_jobs          # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import model_validation_jobs             # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import opportunity_management_jobs       # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import suggestion_stats_computation_jobs # pylint: disable=unused-import  # isort: skip
+from core.jobs.batch_jobs import user_stats_computation_jobs       # pylint: disable=unused-import  # isort: skip
 
 
 def get_all_jobs() -> List[Type[base_jobs.JobBase]]:
