+++ scripts/run_mypy_checks.py
@@ -497,8 +497,8 @@ NOT_FULLY_COVERED_FILES = [
     'extensions/value_generators/models/generators.py',
     'extensions/value_generators/models/generators_test.py',
     'extensions/visualizations/models.py',
-    'core/jobs/batch_jobs/validation_jobs.py',
-    'core/jobs/batch_jobs/validation_jobs_test.py',
+    'core/jobs/batch_jobs/model_validation_jobs.py',
+    'core/jobs/batch_jobs/model_validation_jobs_test.py',
     'core/jobs/types/blog_validation_errors.py',
     'core/jobs/types/blog_validation_errors_test.py',
     'core/jobs/batch_jobs/blog_validation_jobs.py',
